Rework noauth
