+++ plugins/contents/fps_contents/routes.py
@@ -9,7 +9,7 @@ from fps.hooks import register_router  # type: ignore
 from fastapi import APIRouter, Depends
 from starlette.requests import Request  # type: ignore
 
-from fps_auth.routes import current_user  # type: ignore
+from fps_auth.backends import current_user  # type: ignore
 from fps_auth.models import User  # type: ignore
 
 from .models import Checkpoint, Content, SaveContent
@@ -23,7 +23,7 @@ router = APIRouter()
 )
 async def create_content(
     request: Request,
-    user: User = Depends(current_user()),
+    user: User = Depends(current_user),
 ):
     create_content = await request.json()
     path = Path(create_content["path"])
@@ -53,13 +53,13 @@ async def create_content(
 @router.get("/api/contents")
 async def get_root_content(
     content: int,
-    user: User = Depends(current_user()),
+    user: User = Depends(current_user),
 ):
     return Content(**get_path_content(Path(""), bool(content)))
 
 
 @router.get("/api/contents/{path:path}/checkpoints")
-async def get_checkpoint(path, user: User = Depends(current_user())):
+async def get_checkpoint(path, user: User = Depends(current_user)):
     src_path = Path(path)
     dst_path = (
         Path(".ipynb_checkpoints") / f"{src_path.stem}-checkpoint{src_path.suffix}"
@@ -74,7 +74,7 @@ async def get_checkpoint(path, user: User = Depends(current_user())):
 async def get_content(
     path: str,
     content: int,
-    user: User = Depends(current_user()),
+    user: User = Depends(current_user),
 ):
     return Content(**get_path_content(Path(path), bool(content)))
 
@@ -82,7 +82,7 @@ async def get_content(
 @router.put("/api/contents/{path:path}")
 async def save_content(
     request: Request,
-    user: User = Depends(current_user()),
+    user: User = Depends(current_user),
 ):
     save_content = SaveContent(**(await request.json()))
     try:
@@ -110,7 +110,7 @@ async def save_content(
     "/api/contents/{path:path}/checkpoints",
     status_code=201,
 )
-async def create_checkpoint(path, user: User = Depends(current_user())):
+async def create_checkpoint(path, user: User = Depends(current_user)):
     src_path = Path(path)
     dst_path = (
         Path(".ipynb_checkpoints") / f"{src_path.stem}-checkpoint{src_path.suffix}"
